apps/backend/requirements.txt
absl-py==2.3.1
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
aistudio-sdk==0.3.6
alabaster==1.0.0
altair==5.5.0
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==3.7.1
astunparse==1.6.3
asyncio==3.4.3
attrs==25.3.0
babel==2.17.0
backports.tarfile==1.2.0
bce-python-sdk==0.9.45
bcrypt==3.2.2
beanie==1.23.6
billiard==4.2.1
black==23.11.0
blinker==1.9.0
boto3==1.34.43
botocore==1.34.162
Brotli==1.1.0
build==1.3.0
CacheControl==0.14.3
cachetools==5.5.2
celery==5.3.6
celery-sqlalchemy-scheduler==0.3.0
certifi==2025.7.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
cleo==2.1.0
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
coloredlogs==15.0.1
colorlog==6.9.0
contourpy==1.3.2
coverage==7.13.1
crashtest==0.4.1
cryptography==45.0.5
cycler==0.12.1
dataclasses-json==0.6.7
Deprecated==1.2.18
distlib==0.4.0
distro==1.9.0
dnspython==2.7.0
docutils==0.21.2
dulwich==0.24.2
easyocr==1.7.2
ecdsa==0.19.1
email-validator==2.1.0
et_xmlfile==2.0.0
faiss-cpu==1.11.0.post1
fastapi==0.122.0
fastjsonschema==2.21.2
ffmpy==0.6.1
filelock==3.18.0
filetype==1.2.0
findpython==0.7.0
firebase-admin==6.5.0
flatbuffers==25.2.10
flower==2.0.1
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2025.7.0
future==1.0.0
gast==0.6.0
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.7.0
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.2.0
google-crc32c==1.7.1
google-generativeai==0.8.5
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
gradio==5.44.1
gradio_client==1.12.1
groovy==0.1.2
grpcio==1.73.1
grpcio-status==1.71.2
h11==0.16.0
h5py==3.14.0
hf-xet==1.1.5
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.4
humanfriendly==10.0
humanize==4.12.3
idna==3.10
imageio==2.37.0
imagesize==1.4.1
importlib_metadata==8.7.0
indic-nlp-library==0.81
indicnlp==0.0.1
iniconfig==2.1.0
installer==0.7.0
isort==5.12.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
Jinja2==3.1.6
jiter==0.12.0
jmespath==1.0.1
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
keras==3.10.0
keyring==25.6.0
kiwisolver==1.4.8
kombu==5.5.4
langchain==0.2.17
langchain-community==0.2.19
langchain-core==0.3.78
langchain-google-genai==2.1.12
langchain-text-splitters==0.2.4
langsmith==0.4.32
lazy-model==0.2.0
lazy_loader==0.4
libclang==18.1.1
lightgbm==4.6.0
limits==5.4.0
lxml==6.0.0
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
mdurl==0.1.2
ml_dtypes==0.5.1
modelscope==1.29.2
more-itertools==10.8.0
Morfessor==2.0.6
motor==3.6.0
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.3
mypy_extensions==1.1.0
namex==0.1.0
narwhals==2.3.0
networkx==3.5
ninja==1.13.0
numpy==1.26.4
onnxruntime==1.22.1
openai==2.8.0
opencv-contrib-python==4.10.0.84
opencv-python==4.11.0.86
opencv-python-headless==4.9.0.80
openpyxl==3.1.5
opt-einsum==3.3.0
optional-django==0.3.0
optree==0.16.0
orjson==3.11.0
packaging==24.2
paddleocr==3.2.0
paddlepaddle==3.1.1
paddlex==3.2.1
pandas==2.3.1
passlib==1.7.4
pathspec==0.12.1
pbs-installer==2025.10.10
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.10.3
pillow==11.1.0
pinecone==7.3.0
pinecone-plugin-assistant==1.8.0
pinecone-plugin-interface==0.0.7
pkginfo==1.12.1.2
platformdirs==4.4.0
pluggy==1.6.0
poetry==2.2.1
poetry-core==2.2.1
poppler-utils==0.1.0
prettytable==3.16.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.12.5
pydantic-settings==2.1.0
pydantic_core==2.41.5
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.9.1
PyMuPDF==1.23.6
PyMuPDFb==1.23.6
pyparsing==3.2.3
pypdf==3.16.2
pypdfium2==4.30.0
pyproject_hooks==1.2.0
pyspellchecker==0.8.1
pytesseract==0.3.10
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
python-bidi==0.6.6
python-dateutil==2.9.0
python-docx==1.1.0
python-dotenv==1.2.1
python-jose==3.3.0
python-magic==0.4.27
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
qrcode==8.0
RapidFuzz==3.14.1
react==4.3.0
redis==5.0.1
referencing==0.36.2
regex==2023.10.3
reportlab==4.2.0
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.0.0
roman-numerals-py==3.1.0
rpds-py==0.27.1
rsa==4.9.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
ruff==0.12.11
s3transfer==0.10.4
safehttpx==0.1.6
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.16.0
semantic-version==2.10.0
sentence-transformers==5.0.0
sentencepiece==0.2.1
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
slowapi==0.1.9
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==3.0.1
Sphinx==8.2.3
sphinx-argparse==0.5.2
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.23
starlette==0.50.0
streamlit==1.49.1
sympy==1.14.0
tenacity==8.5.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
tf_keras==2.19.0
threadpoolctl==3.6.0
tifffile==2025.8.28
tiktoken==0.12.0
tokenizers==0.22.0
toml==0.10.2
tomlkit==0.13.3
torch==2.7.1
torchvision==0.22.1
tornado==6.5.1
tqdm==4.67.1
transformers==4.56.1
trove-classifiers==2025.9.11.17
typer==0.17.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.14.1
tzdata==2025.2
ujson==5.11.0
ultralytics==8.3.169
ultralytics-thop==2.0.14
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.21.0
vine==5.1.0
virtualenv==20.35.3
watchfiles==1.1.0
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.1.3
wrapt==1.17.2
xattr==1.2.0
xgboost==3.0.5
xlrd==2.0.2
yarl==1.20.1
zipp==3.23.0
zstandard==0.25.0
